product delete is possible

+++ index.js
@@ -8,7 +8,7 @@ const port = process.env.PORT || 5000;
 app.use(cors());
 app.use(express.json());
 
-const { MongoClient, ServerApiVersion, Admin } = require("mongodb");
+const { MongoClient, ServerApiVersion, Admin, ObjectId } = require("mongodb");
 const uri = `mongodb+srv://admin:${process.env.PASSWORD}@database.mp0iy.mongodb.net/?retryWrites=true&w=majority`;
 const client = new MongoClient(uri, {
   useNewUrlParser: true,
@@ -24,11 +24,33 @@ const varifyJWT = (req, res, next) => {
   req.token = token;
   next();
 };
+
 const run = async () => {
   try {
     await client.connect();
     console.log("connected with mongodb");
     const userCollection = client.db("users").collection("user");
+    const productCollection = client.db("products").collection("product");
+
+    // for product
+    app.get("/product", async (req, res) => {
+      const query = {};
+      const result = await productCollection.find(query).toArray();
+      res.send(result);
+    });
+    app.post("/product", async (req, res) => {
+      const product = req.body;
+      const result = await productCollection.insertOne(product);
+      res.send(result);
+    });
+    app.delete("/product/:id", async (req, res) => {
+      const { id } = req.params;
+      const query = { _id: ObjectId(id) };
+      const result = await productCollection.deleteOne(query);
+      res.send(result);
+    });
+    //for users
+    // find all users
     app.get("/users", varifyJWT, async (req, res) => {
       if (req.token.role === "admin") {
         const users = await userCollection.find().toArray();
@@ -37,8 +59,8 @@ const run = async () => {
         res.send({ error: "unauthorised acces" });
       }
     });
+    // validation of admin
     app.get("/user/admin/:email", varifyJWT, async (req, res) => {
-      console.log(req.token);
       const { email } = req.params;
       const decodedEmail = req.token.email;
       if (email === decodedEmail) {
@@ -49,6 +71,7 @@ const run = async () => {
         res.send({ error: "forbidden" });
       }
     });
+    // to make a user admin
     app.put("/user/admin/:email", async (req, res) => {
       const { email } = req.params;
       const filter = { email };
@@ -59,15 +82,19 @@ const run = async () => {
         },
       };
       const result = await userCollection.updateOne(filter, updateDoc, options);
-      console.log(result);
       res.send({ result: result });
     });
-
-    app.delete("/user/:email", async (req, res) => {
-      const { email } = req.params;
-      const query = { email };
-      const result = userCollection.deleteOne(query);
-      res.send(result);
+    //deleting a user
+    app.delete("/user/:email", varifyJWT, async (req, res) => {
+      const { token } = req;
+      if (token.role === "admin") {
+        const { email } = req.params;
+        const query = { email };
+        const result = userCollection.deleteOne(query);
+        res.send(result);
+      } else {
+        res.send({ error: "unauthorised" });
+      }
     });
     // this can handle any kind of login
     app.post("/user/:email", async (req, res) => {
